Fixed typos in README, fixed unescaped ampersand in results.
Device responses can include an ampersand, eg to combine the
names of two artists. Correct XML would have escaped the ampersand
character, but the device does not adhere to the XML standard.
This is now fixed by escaping the ampersand just before the
text is being decoded as XML.

The tests.py file is updated to fix pep and flake issues.